--- a/Lib/test/test_deque.py
+++ b/Lib/test/test_deque.py
@@ -518,7 +518,7 @@ class TestBasic(unittest.TestCase):
         for match in (True, False):
             d = deque(['ab'])
             d.extend([MutateCmp(d, match), 'c'])
-            self.assertRaises(IndexError, d.remove, 'c')
+            self.assertRaises(RuntimeError, d.remove, 'c')
             self.assertEqual(d, deque())
 
     def test_repr(self):