--- a/Misc/NEWS.d/next/Core and Builtins/2018-06-13-00-23-19.bpo-25246.HKOrh3.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2018-06-13-00-23-19.bpo-25246.HKOrh3.rst
@@ -1,0 +1,2 @@
+Improve the performance of :meth:`collections.deque.remove` and raise
+`exc`:`RuntimeError` if the deque is mutated during removal.