❯ ./python -m perf compare_to old.json .new.json
Mean +- std dev: [old] 214 us +- 5 us -> [new] 147 us +- 2 us: 1.46x faster (-31%)


<!-- issue-number: [bpo-25246](https://bugs.python.org/issue25246) -->
https://bugs.python.org/issue25246
<!-- /issue-number -->