@pablogsal 2018-10-13
@taleinat I created this new PR as the old patch did not cleanly apply to the current master and I wanted to re-run the performance benchmarks again. This PR uses the same strategy as in the old one as other alternatives that I tried (like reducing branch mispredictions or inlining some parts of the code) did not show much better performance improvements.