@pablogsal 2018-10-13
Here are some extra benchmarks that I took:

BASELINE
```
 Performance counter stats for './python -c
import collections
D = collections.deque(range(100000))
for _ in range(100):
    D.remove(100000//2 + _)
' (200 runs):

         1,390,689      cache-references:u                                            ( +-  0.56% )
           134,664      cache-misses:u            #    9.683 % of all cache refs      ( +-  2.84% )
       518,290,820      cycles:u                                                      ( +-  0.39% )
     1,573,923,571      instructions:u            #    3.04  insn per cycle           ( +-  0.00% )
       373,948,261      branches:u                                                    ( +-  0.00% )
             3,353      faults:u                                                      ( +-  0.00% )
                 0      migrations:u

          0.161307 +- 0.000670 seconds time elapsed  ( +-  0.42% )
```

PATCH
```
 Performance counter stats for './python -c
import collections
D = collections.deque(range(100000))
for _ in range(100):
    D.remove(100000//2 + _)
' (200 runs):

         1,506,197      cache-references:u                                            ( +-  0.43% )
           131,032      cache-misses:u            #    8.699 % of all cache refs      ( +-  2.65% )
       402,143,102      cycles:u                                                      ( +-  0.52% )
     1,135,753,509      instructions:u            #    2.82  insn per cycle           ( +-  0.00% )
       289,432,604      branches:u                                                    ( +-  0.00% )
             3,351      faults:u                                                      ( +-  0.00% )
                 0      migrations:u

          0.126766 +- 0.000642 seconds time elapsed  ( +-  0.51% )
```