@taleinat 2018-10-13
@pablogsal

> If you prefer, I can reopen the old one again instead of using this one :)

In this case let it be, but in the future it would be better to re-open such a PR rather than create a new one.