@taleinat 2018-10-13
> I refactored the logic into `_deque_index` because the code is pretty much the same for `deque_index` and `deque_remove`.

Sorry, I missed that!

> Should I create a separate PR if this gets merged for refactoring `deque_index` and `deque_remove`?

No.