@pablogsal 2018-10-13
Basically, this PR is:

* A more performant deque_index that can be reusable (now `_deque_index`).
* Change `deque_remove` to be: 
  1. Find index of element in deque.
  2. Use `deque_del_item` with that index.
* Refactor `deque_index` to use the `_deque_index` function.