@pablogsal 2018-10-13
More benchmarks. Trying to remove an element that is not in the deque:
```python
D = collections.deque(range(19999))
def test_func():
    try:
        D.remove("f")
    except:
        pass
```


❯ ./python -m perf compare_to old.json ../cpython/new.json
Mean +- std dev: [old] 483 us +- 16 us -> [new] 349 us +- 6 us: 1.38x faster (-28%)