@rhettinger 2018-10-14
Thanks for the patch.  At first glance the code looks good and I've verified the performance boost for remove().  Will take a more thorough look later (it may take a while).

Reminders for myself on what to think about:

* Misgivings about the *RuntimeError* or *IndexError*. This may be difficult to avoid in multithreaded code. Also, it impairs substitutability with lists -- one of the principal goals for adding *remove()* was better substitutability for lists.  Is there a way to drop the error and do something closer to what lists would do?

* For most practical uses for *deque.remove()*, the running time is dominated by the rich comparison equality tests for a bunch of objects randomly located in memory.  This PR doesn't help (and can't help) those cases (i.e. removing scheduled tasks from a deque used as a FIFO queue).

* Compare the perf improvement with an alternative patch that implements an inline "_deque_rotate(deque, -1)".  This may give a similar performance improvement.

* The new index/delitem logic is more natural than the rotate/del/rotate logic.  At first glance, it should have better cache performance -- think this through to make sure.