@rhettinger 2019-01-06
Here's a little more timing analysis after this PR is applied.  For a thousand iterations, the new *remove()* takes 18.8ms.  That is very close to the sum of indexing alone at 17.5ms and plus deletion alone at 1.4ms.  Both of latter timings are substantially unchanged by the PR.

This confirms the intuition that once this patch is applied, *remove()* performance is 93% dominated by the indexing step.  That means that *remove()* will be very close to as good as it is ever going to get (i.e. if the deletion step were improved by 50%, *remove()* would only improve by 3%).

```
$ ./python.exe -m timeit -r11 -s 'from collections import deque' -s 'd0 = deque(range(4_010))' -s 'r = range(2_000, 1_000, -1)' 'd=d0.copy()' 'm = d.remove'  'for i in r: m(i)'
20 loops, best of 11: 18.8 msec per loop

$ ./python.exe -m timeit -r11 -s 'from collections import deque' -s 'd0 = deque(range(4_010))' -s 'r = range(2_000, 1_000, -1)' 'd=d0.copy()' 'm = d.remove'  'for i in r: del d[i]'
200 loops, best of 11: 1.42 msec per loop

$ ./python.exe -m timeit -r11 -s 'from collections import deque' -s 'd0 = deque(range(4_010))' -s 'r = range(2_000, 1_000, -1)' 'd=d0.copy()' 'm = d.index'  'for i in r: m(i)'
20 loops, best of 11: 17.4 msec per loop

$ ./python.exe -m timeit -r11 -s 'from collections import deque' -s 'd0 = deque(range(4_010))' -s 'r = range(2_000, 1_000, -1)' 'd=d0.copy()' 'm = d.index'  'for i in r: pass'
5000 loops, best of 11: 44.3 usec per loop
```