@rhettinger 2019-01-07
I'm still thinking about the error reporting in the face of mutation.  

The *remove()* method in lists makes no effort to detect or report mutation; instead, it just checks to make sure it doesn't index past the then current end of the list.  If the value is found, an attempt is made to remove it. There is no checking to make sure it hasn't moved.  Any size changes are ignored.

For deques, we can't just keep iterating once the deque has mutated.  We could start over (similar to the strategy that dict/set lookup uses), or we could stop looping and raise ValueError (similar to what dicts do).

We could keep the API the same as we have now (reporting an IndexError instead of a RuntimeError).  That would avoid breaking any code (probably rare) that relies on the current behavior.

Do you have any thoughts on what would be best for the users?