@pablogsal 2019-01-09
> Do you have any thoughts on what would be best for the users?

In my opinion, we should try to inform the users with an informative exception (IndexError for example) in case the deque is mutated. The reason is that we should not encourage mutating operations during iteration as those tend to be very dangerous. Even if in some cases it results in what the user wants, _errors should not pass silently_. This also has the advantage of being consistent with other containers and to be slightly more backwards compatible. This allows the user also to act exactly as they want if they indeed intend to mutate the deque in these cases.

Also, I think starting over in a method that mutates the container (as opposed to some that not as looking up something in dics/sets) is also dangerous and substantially more error prone.